Move SplitBytes function to common module
