+++ prover/zkevm/prover/common/split_test.go
@@ -1,4 +1,4 @@
-package ecdsa
+package common
 
 import (
 	"github.com/stretchr/testify/assert"
