+++ prover/zkevm/prover/ecdsa/unaligned_gnark.go
@@ -185,7 +185,7 @@ func (d *UnalignedGnarkData) assignUnalignedGnarkData(run *wizard.ProverRuntime,
 			var txBts []byte
 			for _, txHighBtsRow := range rows[4*common.NbLimbU128 : 6*common.NbLimbU128] {
 				bytes := txHighBtsRow.Bytes()
-				txBts = append(txBts, bytes[field.Bytes-LimbBytes:]...)
+				txBts = append(txBts, bytes[field.Bytes-common.LimbBytes:]...)
 			}
 
 			copy(prehashedMsg[:], txBts[:])
@@ -193,7 +193,7 @@ func (d *UnalignedGnarkData) assignUnalignedGnarkData(run *wizard.ProverRuntime,
 			var vBts []byte
 			for _, vBtsRow := range rows[6*common.NbLimbU128 : 8*common.NbLimbU128] {
 				bytes := vBtsRow.Bytes()
-				vBts = append(vBts, bytes[field.Bytes-LimbBytes:]...)
+				vBts = append(vBts, bytes[field.Bytes-common.LimbBytes:]...)
 			}
 
 			v.SetBytes(vBts)
@@ -201,7 +201,7 @@ func (d *UnalignedGnarkData) assignUnalignedGnarkData(run *wizard.ProverRuntime,
 			var rBts []byte
 			for _, rBtsRow := range rows[8*common.NbLimbU128 : 10*common.NbLimbU128] {
 				bytes := rBtsRow.Bytes()
-				rBts = append(rBts, bytes[field.Bytes-LimbBytes:]...)
+				rBts = append(rBts, bytes[field.Bytes-common.LimbBytes:]...)
 			}
 
 			r.SetBytes(rBts)
@@ -209,7 +209,7 @@ func (d *UnalignedGnarkData) assignUnalignedGnarkData(run *wizard.ProverRuntime,
 			var sBts []byte
 			for _, sBtsRow := range rows[10*common.NbLimbU128 : 12*common.NbLimbU128] {
 				bytes := sBtsRow.Bytes()
-				sBts = append(sBts, bytes[field.Bytes-LimbBytes:]...)
+				sBts = append(sBts, bytes[field.Bytes-common.LimbBytes:]...)
 			}
 
 			s.SetBytes(sBts)
@@ -238,7 +238,7 @@ func (d *UnalignedGnarkData) assignUnalignedGnarkData(run *wizard.ProverRuntime,
 			var txBts []byte
 			for _, txHighBtsRow := range rows[4*common.NbLimbU128 : 6*common.NbLimbU128] {
 				bytes := txHighBtsRow.Bytes()
-				txBts = append(txBts, bytes[field.Bytes-LimbBytes:]...)
+				txBts = append(txBts, bytes[field.Bytes-common.LimbBytes:]...)
 			}
 
 			copy(prehashedMsg[:], txBts[:])
@@ -249,21 +249,21 @@ func (d *UnalignedGnarkData) assignUnalignedGnarkData(run *wizard.ProverRuntime,
 			}
 			v.FillBytes(buf[:])
 
-			vLimbs := SplitBytes(buf[:])
+			vLimbs := common.SplitBytes(buf[:])
 			for j, vLimb := range vLimbs {
 				rows[6*common.NbLimbU128+j].SetBytes(vLimb)
 			}
 
 			r.FillBytes(buf[:])
 
-			rLimbs := SplitBytes(buf[:])
+			rLimbs := common.SplitBytes(buf[:])
 			for j, rLimb := range rLimbs {
 				rows[8*common.NbLimbU128+j].SetBytes(rLimb)
 			}
 
 			s.FillBytes(buf[:])
 
-			sLimbs := SplitBytes(buf[:])
+			sLimbs := common.SplitBytes(buf[:])
 			for j, sLimb := range sLimbs {
 				rows[10*common.NbLimbU128+j].SetBytes(sLimb)
 			}
@@ -284,13 +284,13 @@ func (d *UnalignedGnarkData) assignUnalignedGnarkData(run *wizard.ProverRuntime,
 			utils.Panic("error recovering public: err=%v v=%v r=%v s=%v", err.Error(), v.Uint64()-27, r.String(), s.String())
 		}
 		pkx := pk.A.X.Bytes()
-		pkxLimbs := SplitBytes(pkx[:])
+		pkxLimbs := common.SplitBytes(pkx[:])
 		for j, xLimb := range pkxLimbs {
 			rows[j].SetBytes(xLimb)
 		}
 
 		pky := pk.A.Y.Bytes()
-		pkyLimbs := SplitBytes(pky[:])
+		pkyLimbs := common.SplitBytes(pky[:])
 		for j, yLimb := range pkyLimbs {
 			rows[2*common.NbLimbU128+j].SetBytes(yLimb)
 		}
@@ -320,7 +320,7 @@ func (d *UnalignedGnarkData) assignUnalignedGnarkData(run *wizard.ProverRuntime,
 		for j := 0; j < nbRowsPerGnarkPushing*common.NbLimbU128; j++ {
 			rowBytes := rows[j].Bytes()
 			bytes := [16]byte{}
-			copy(bytes[:], rowBytes[field.Bytes-LimbBytes:])
+			copy(bytes[:], rowBytes[field.Bytes-common.LimbBytes:])
 
 			var elementLA field.Element
 			elementLA.SetBytes(bytes[:])
